Updated ICombat to use an int for number of frames stunned rather than a double based on time

+++ ActualGame/ActualGame/ICombat.cs
@@ -15,6 +15,6 @@ namespace ActualGame
         void Die();
 
         //A method to stun a Character for a specific amount of time if they take damage
-        void Stun(double timeToStun);
+        void Stun(int timeToStun);
     }
 }
